Refactor metrics classes
Regenerate ConscryptStatsLog based on the latest proto definition. This
class is nearly identical to the generated wrapper except for the usage
of ReflexiveStatsEvent and ReflexiveStatsLog. Updated the class
documentation with the constraints around its usage.

buildEvent methods are not used as non-existent in the wrapper. They are
marked @Deprecated but not removed because implicitly used by
MetricsTest in CTS. ReflexiveStatsEvent builder is updated to ignore any
call to writeIntArray if not supported by the platform (i.e., sdk < 32).

The Source class is updated to use the ConscryptStatsLog constants. A
new class CertificateTransparencyVerificationReason is added to
encapsulate the reason constants.

From a behaviour's perspective, this change is a no-op.

Test: atest
Bug: 319829948
Change-Id: Ie6b166530544345491347ff6bda8bbf4633e37c3