Remove from/to attributes from libcore.util.Nullable/NonNull
The attributes are not used, either in the code, or in Metalava and
their presence prevents mapping them to other, more common nullability
annotations like those in `android.annotation` or `androidx.annotation`
as Metalava needs to do when generating the stubs.

Matches the changes made in https://r.android.com/3508044.

Bug: 311619997
Test: m checkbuild
Change-Id: Icf0ee4fa4c6510cdd2a2c9921ef333fd568f1a91