CT: Support static-ct-api logs
Update the policy to support Static API logs as described in the
announcement email [1]. Add an enum to LogInfo to capture the type of
log (TYPE_RFC6962 or TYPE_STATIC). Add NO_RFC6962_LOG to
PolicyCompliance enum to capture the new failure.

Bump the compatibility version to 2.

Remove the URL field from the LogInfo (not used).

[1] https://groups.google.com/a/chromium.org/g/ct-policy/c/HBFZHG0TCsY

Bug: 377276991
Test: atest ConscryptPrivateTestCases
Change-Id: Id8077646816451d1b75407e029765dd40e9b3f93